fetch asset list from API and allow deletion
Use redux-api-middleware to allow fetching assets from the API and
deleting them.

Separate out the AssetList table into a separate AssetTable container
which has the table headings encapsulated within it.

Connect up components to the redux state in order to provide
functionality. This makes the AssetList component particularly compact.

Currently the API is hard-wired to be at http://localhost:8080/.

Closes #12.