update OAuth2 callback URL
It's a bit messy having the OAuth2 callback URL be an active page of the
app. It also introduces a small race between redux-implicit-oauth2
listening for the callback's hash and that callback being redirected to
the login page.

Fix both of these by having a dedicated, empty, non-redirected endpoint
for oauth2 callbacks.