add authorisation headers to API requests
Add some custom middleware which takes the token from the redux state
and adds it to the authorisation headers of the HTTP response. For the
moment we simply log the user out if the token is invalid or times out.
For a more sophisticated UI, we might want to look at
https://github.com/esbenp/redux-refresh-token to automate the refresh of
the token.