add authorisation headers to API requests
Add redux-api-middleware which interprets specially formatted "RSAA"
actions as calls to a HTTP API. Add some custom middleware which takes
the token from the redux state and adds it to the authorisation headers
of this HTTP response. For the moment we simply log the user out if the
token is invalid or times out.

For a more sophisticated UI, we might want to look at
https://github.com/esbenp/redux-refresh-token to automate the refresh of
the token. This will require support in the consent app to provide an
"early fail" UI or "is UI-less login possible" endpoint in order to
provide a seamless experience.

Closes #25.