assetRegisterApi: add getAssets action which queries the asset list
Add action which can take sort, filtering and search parameters which
are passed to the asset API endpoint. Update the assets state with a
query object which records the sort, filter and search parameters used.

The default ordering is by descending updated at time. (I.e., the
"freshest" objects are first.)