add help content for static page
Since the re-design there is now only one static page. (And possibly
"Feedback".) Lift the content from the document cited in #11 and add
some basic styling for the page.

Static page content and titles are now defined in ES6 modules under
src/static for convenience of editing. The Static component is re-worked
to accept a page name corresponding to the module name under src/static.

Update the sidebar to link to the only static page we currently have.

Closes #11