add basic Sidebar styling
Style the sidebar to be correctly positioned and sized. Arrange the menu
as per the IAR design. The navigation is implemented via the Material UI
List component but with a custom SidebarNavLink component which renders
a link to static pages and styles itself appropriately when the currentl
location matches the link.