reset asset list summaries on successful PUT/POST
If an asset is created or modified via POST or PUT, we cannot in general
in the frontend determine if it should appear in the current summary
list without essentially replicating the sorting, filtering and search
functionality of the backend.

For the moment, take the correct but, perhaps, inelegant approach of
resetting our cached list of asset summaries when an asset is created or
modified. This will cause the asset list to be re-fetched when the asset
list table is next shown.