asset list: display institutions using their lookup name
Add a list of institutions keyed by instid to the redux store and add a
new action, listInstitutions(), which will populate it.

Add a LookupInstitution component which renders a lookup instution given
the lookup instid falling back to using the instid.

The initial fetch is triggered by a FetchLookupInstitutions component
which will cause a fetch of institutions on first render if a fetch had
not previous occurred. This component is rendered outside of the router
component since it should only fetch on first page load, not on
navigation.