reset all application state on logout
LOGOUT is a special action in that potentially all state cached by the
application becomes invalid.

Wrap the default reducer in an implementation which resets state by
calling the reducers with undefined as the current state. Reducers are
given a change to "save" something from the original state since it is
added to the LOGOUT action as an originalState property but a "naive"
reducer will just return the initial state for the application.