re-style AssetForm
This is work-in-progress refactor of the AssetForm component. Since it
is a large component, we break it out into sub-components within the
src/AssetForm/ directory.

The AssetForm consists of two classes of component: presentational and
containers. Presentational components are pure functions of their input
props whereas container components wrap presentational components and
set props depending on application state.

Most of the styling work has been done by breaking out the asset form
into presentational components for the header and body and further
breaking the body into different sections. This should make each
component small enough to fit in a human's head :).

The container components provide information about the current draft to
the presentational components. Aside from some small containers which
fetch a draft, check for completeness, etc, the bulk of the form is
is implemented via <DraftInput>.