re-work delete confirmation to return promise
Dispatching the confirmDelete action now returns a Promise which is
resolved/rejected with the URL depending on whether the user
confirms/cancels the deletion.

The reasoning for this is that the delete asset button on the asset edit
form needs to wait for the confirmation to go through before deciding
whether to navigate away from the page.