NavigateBackButton: generalise to wrap any onClick-able component
NavigateBackButton's functionality is of general utility. Generalise it
to set the onClick prop of a wrapped component which can be overridden
via the "component" prop. Default functionality remains the same.

The motivation for this is handling the "cancel" button on the edit
asset form which has essentially the same functionality as
NavigateBackButton.