add withScroll higher order component
Add a higher order component which subscribes to scroll events on the
current window. Takes a single function, mapScrollToProps, which is
passed the scroll events as they arrive and returns an object of props
to set on the wrapped component.

This is required to style the asset form header so that it develops a
shadow as the user begins to scroll the page.

Example of use:

const mapScrollToProps = ({ pageY }) => ({ hasScrolled: pageY > 0 });

const SensitiveWidget = withScroll(mapScrollToProps)(
  ({ hasScrolled }) => <span>
    has scrolled: { hasScrolled ? 'yes' : 'no' }
  </span>
)