add module which provides support for editing the current draft
Provide a new module, "draft", which can be used to access and edit the
current draft.

The core of this module is the "withDraft" higher order component which
maps the current draft and functions which can modify it into props to
be set on the component.

It takes a single argument, mapDraftToProps, which is a function called
with three arguments:

  - an object representing the current draft state of the form
    {
      draft: <asset resource being edited>
      isLoading: boolean indicating if a new draft is being loaded from
        an existing asset
      isLive: boolean indicating if the current draft is "live" (has
        been loaded or created and not yet saved)
    }
  - the wrapped component's own props
  - an "editor" object which has three function properties:
    fetchOrCreateDraft, patchDraft and saveDraft. See these actions'
    documentation for details.

The mapDraftToProps function should return an object of props to set on
the wrapped component.

The rationale for gating everything through the withDraft() HoC is that
the fact that the draft is currently implemented as redux store state
limits us to one draft per application. This almost certainly will
change in future so having the withDraft() HoC allows us the exit route
of moving to a <DraftProvider> like component which sets the draft on
the current context.

Time limitations prevent the implementation of <DraftProvider> in the
current sprint.

In addition to withDraft(), the DraftInput component is provided to
allow <input>-like components to edit the current draft.

DraftInput is a container component which wraps an input-like component
and takes care of setting value and handling onChange to connect the
component to the current asset draft. This component takes a couple of
mapping functions to move from draft to input props and on change events
to draft patches.

A few general-use specialisations of the component are provided in
DraftInput.js.