src/index.js: un register the service worker
As noted in [1], the service worker we register in index.js provides an
additional level of offline caching for the application. This provides
very rapid load times on, e.g., Google Chrome but has the downside that
updates to the application may take 24 hours to propagate.

While we still may make lock-step backend/frontend changes, disable and
un-register (if registered) the service worker. We can re-visit this
once the application stabilises.

Closes #91.

[1]
https://github.com/superhighfives/create-react-app/blob/daaa9b2c79ffa9fc4a81a28ca2fd15872f3aa50a/packages/react-scripts/template/README.md#opting-out-of-caching