add new container: WaitForGlobals
Add WaitForGlobals component which will only render its children once
the user is logged in, their profile has been retrieved and the list of
institutions is available.

This will be used for the asset list, edit form and read-only view to
avoid flashes of partial data as we wait for these global values to
become available.

The component renders a full-screen loading indicator page which is
gracefully hidden when resources are loaded.