add read-only view of asset record
Add a read-only variant of AssetForm called AssetView. Move the existing
src/AssetForm directory to src/AssetDetail since it now contains more
than just the form.

Most of the presentational components like the header, section bars, etc
are re-used from the AssetForm. We add some custom read-only input-like
fields to provide Material UI input styling for the read only view. This
is a little bit of a HACK but provides some consistency between the
read-only and edit asset views.

The URL scheme has been changed to have /asset/[...]/edit be a link to
the edit form and /asset/[...] a link to the read-only view.
AssetListItem has been taught how to link to the correct view depending
on the user's permissions.

The rationale for still allowing a user to visit the .../edit link even
if they do not appear to have permissions is to allow admin users to
edit assets. Currently only the backend knows who the admin users are
and so the frontend has no way of performing this check automatically.
We assume that admin users are sufficiently sophisticated to construct
their own .../edit URLs and should a user come across .../edit
inadvertently, the backend will still stop them being able to edit
records.