add new component: PreventNavigationWhenDraftModified
Add a new component, PreventNavigationWhenDraftModified, which examines
the isModified flag in the current draft and prompts for confirmation
when navigating away from the page if isModified is set to true.

This change exposed an issue in the use of history.goBack() to implement
the save and cancel buttons as outlined in #104. Addressing #104 is not
viewed part of the story for this commit, however.