IE11: use polyfill.io and browser-update.org.
Polyfill.io provide a service which provides a minified set of polyfills
tailored to your browser. We're unlikely to do any better than their
implementations, so make use of them.

Browser-update.org provides a canned "please upgrade your browser"
message which will be displayed once to IE11 users. Add this in to at
least warn users that their experience is (very) likely to be degraded.