HACK: allow searching through "secret" ?q=[...] URL
This is a little hack to expose the search functionality by a "hidden"
URL parameter on the asset entry list pages.

If "?q=[...]" is appended to the URL, the value is passed to the backend
as the "search" parameter.

This is intended a) to make it easy to evaluate the search functionality
and b) to aid UX in finding and evaluating usages of the IAR.