add theme overrides for controls
Restore the original colours of form components and links in form
control helper text by means of a custom override in the MUI theme. This
is a little bit of a hack in that we are, effectively, forcing
"colour='secondary'" to mean "color='primary'". However, fixing up each
use of the control in the entire project is troublesome, error prone and
brittle.