disable Save Entry button if department is blank
If a user with two or more institutions creates a new entry, the
"institution" field is initially blank. Trying to save the entry will
result in a HTTP 400 error.

The underlying reason is that the department field must not be blank.

Ensure that the Save Entry button is disabled until the department field
is non-blank.