assetRegisterApi: add new state value: isExtendingSummaries
Add a new boolean value to the assetRegisterApi state,
isExtendingSummaries, which indicates if the current request will extend
the summary list as opposed to replacing it.

This is intended to be used by the frontend logic to allow the asset
table to be hidden when we're replacing it but not when we're extending
it. (See #146.)