add a basic OAuth2 implicit flow implementation
In order to implement interactive and non-interactive login, we need a
slightly lower-level access to the OAuth2 flow. Move one step down the
chain and use Mulesoft's OAuth2 client to implement implicit flow
ourselves.

This commit contains the implementation of the OAuth2 implicit flow but
does not replace the existing uses of redux-implicit-oauth2.

The new implicit flow implementation also fixes a long-standing niggle
with the existing implementation which would repeatedly poll the window
to see when the OAuth2 callback had been called spamming the error log
as it did so. This new implementation replaces the polling with an
event-driven model using postMessage.

Unfortunately, our existing callback URL was missing the .html extension
which just makes it a little non-trivial to make a like-for-like
replacement in a clean manner. Add a simple static file to
public/oauth2-callback.html which provides the other end of the
postMessage implementation. Update create-client.sh to add this as a
redirect URL.

**IMPORTANT:** this will necessitate changes to the IAR deployment.

A benefit from this is we no-longer need to special case oauth2-callback
in our routing when we replace the existing implementation.