replace existing redux-implicit-oauth2 implementation
Replace the existing redux-implicit-oauth2 authorisation with a
work-alike which uses the same format for the redux state.

This is a little bit of a halfway-horse; ideally we'd store the full
client-oauth2 token object in the state but a lot of the rest of the app
depends on the state having the same form and this commit is intended to
be a small surgical change.

Since we already hid the redux-implicit-oauth2 actions inside the
src/redux/actions/auth.js module, replacing them is relatively
straightforward. We add the ability to do prompt-less login via the
login action. This is currently unused but is in place for the token
timeout handling behaviour.

We keep the existing behaviour of keeping the auth token in localStorage
so that re-visits to the app can have speedy login. We load it as a
by-product of initialising the local state which means we can do away
with the existing auth middleware.

Since we no-longer use the redux-implicit-oauth2 implementation, we can
remove the special-case route in AppRoutes for it.