add token refresh on timeout
Add a new component, TokenTimeout, which maintains an internal timer
which fires just before the access token times out. On time out, attempt
a prompt-less login to refresh the token. If this login fails, log the
user out with an informative message.

There's no link between the consent app's idea of which user is
currently signed in and the IAR's so on a successful prompt-less login,
re-fetch the user's lookup identity and compare it against the identity
of the user who first logged in. If they do not match, log the current
user out with an informative message.

This latter check is only likely to be triggered if another user had
logged into the IAR (or another OAuth2 protected site) within the same
browser context in the interim.