snackbar: do not clear message on close
This commit addresses a small bug found when implementing token timeout
notifications. These notifications are slightly longer than others we've
been using.

Previously the snackbar message was reset when reducing the
SNACKBAR_CLOSE action. This caused a small visual bug because the
snackbar animates shut when closing and the message just blanks as soon
as the animation starts.

This is less visible for short messages, but for long messages the
snackbar visibly "jumps" in size as it closes.

Fix the problem by simply keeping the message as is when reducing the
SNACKBAR_CLOSE action.